EQL: Internal refactoring
Combine Matcher and SequenceStateMachine into one class since the two
have evolved beyond their original name and structure.
